requirements.txt
PySide6>=6.0.0
openai>=1.0.0
python-dotenv>=0.19.0
whisper>=1.0.0
pyaudio
keyboard
pydub
ffmpeg-python
sounddevice
numpy 